Remove unused fields & rename
