+++ toolpad/application.yml
@@ -1,3 +1,3 @@
 apiVersion: v1
 kind: application
-spec: {}
+spec: { authentication: {}, authorization: {} }
