+++ toolpad/pages/page/page.yml
@@ -13,19 +13,6 @@ spec:
         rowsSource: dataProvider
         dataProviderId: dataProvider.ts:default
         columns:
-          - field: id
-            type: number
           - field: name
             type: string
-          - field: account_creation_date
-            type: date
-          - field: country_of_residence
-            type: string
-          - field: phone_number
-            type: number
-          - field: email
-            type: string
-          - field: address
-            type: string
-          - field: gender
-            type: string
+        rowIdField: name
